% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{actually_macro_genre}
\alias{actually_macro_genre}
\title{Frequency data for actually in ICE-GB (macro genres)}
\format{
\subsection{\code{actually_macro_genre}}{

A data frame with 12 rows and 3 columns:
\describe{
\item{macro_genre}{12 macro genres (e.g. "private_dialogues", "student_writing", "reportage"); ordered factor}
\item{actually}{The number of times \emph{actually} occurs in the macro genre}
\item{word_count}{The size of the macros genre, expressed as the totel number of (word and nonword) tokens}
}
Nelson, Gerald, Sean Wallis, and Bas Aarts. 2002. \emph{Exploring Natural Language: Working with the British Component of the International Corpus of English}. John Benjamins.
}
}
\usage{
actually_macro_genre
}
\description{
Data on the number of occurrences of \emph{actually} in each of the 12 macro genres represented in ICE-GB (Nelson et al. 2002).
}
\keyword{datasets}
