% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_actually}
\alias{data_actually}
\title{Actually in contemporary British speech}
\format{
\subsection{\code{data_actually}}{

A data frame with 656 rows and 6 columns:
\describe{
\item{speaker}{speaker ID}
\item{gender}{self-reported gender ("Male", "Female")}
\item{age_group}{age of speaker, 10 binned categories}
\item{n_actually}{number of occurrences of actually}
\item{n_words}{total number of words contributed by the speaker}
\item{rate_pmw}{normalized frequency of actually, per million words}
}
}
}
\source{
Love, Robbie, Claire Dembry, Andrew Hardie, Vaclav Brezina & Tony McEnery. 2017. The Spoken BNC2014: Designing and building a spoken corpus of everyday conversations. \emph{International Journal of Corpus Linguistics}, 22(3), 319--344.

Sönning, Lukas & Manfred Krug. 2021. Actually in contemporary British speech: Data from the Spoken BNC corpora. DataverseNO, V1. https://doi.org/10.18710/A3SATC
}
\usage{
data_actually
}
\description{
Data on the usage rate of actually in spoken British English, drawn from the Spoken BNC2014 (Love et al. 2017).
}
\keyword{datasets}
