% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_adj_gradation}
\alias{data_adj_gradation}
\title{Adjective gradation in 1960s written American English}
\format{
\subsection{\code{data_adj_gradation}}{

A data frame with 817 rows and 9 columns:
\describe{
\item{adjective}{the adjectival base}
\item{form}{"a" analytic (more/most), "s" synthetic (-er/-est)}
\item{category}{"C" comparative, "S" superlative}
\item{text}{text identifier}
\item{genre}{genre of the text (4 categories: "drama", "fiction", "newspaper", "nonfiction")}
\item{position}{syntactic position: "attributive", "predicative", "postnominal", "nominalized"}
\item{than_compl}{than-complement (comparative clause) present? 1 = Yes, 0 = no}
\item{attributive}{whether the adjective occurs in atributive position: 1 = Yes, 0 = no}
\item{coordinated}{do two (or more) graded adjectives occur in coordination? 1 = Yes, 0 = no}
}
}
}
\source{
Braun, Albert. 1982. Studien zu Syntax und Morphologie der Steigerungsformen im Englischen. Bern: Franke.
}
\usage{
data_adj_gradation
}
\description{
Data on the gradation of adjectives in written American English, from Braun (1982). The data only include adjectives that are known to occur with both strategies (inflectional -er/-est and periphrastic more/most). Includes 817 tokens, 139 adjectival types, and information on syntactic context.
}
\keyword{datasets}
