% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_laterals}
\alias{data_laterals}
\title{Laterals in German Learner English}
\format{
\subsection{\code{data_laterals}}{

A data frame with 147 rows and 5 columns:
\describe{
\item{speaker}{speaker ID}
\item{proficiency}{English proficiency of the speaker, expressed as a z-score}
\item{word}{carrier word of the lateral}
\item{context}{phonetic context of the lateral ("cluster", "singleton", "syllabic")}
\item{dark}{realization as dark (1) or clear (0)}
}
}
}
\source{
Pascoe, Graham. 1987. Die Aussprache des Englischen an bayerischen Schulen. München: Profil.
}
\usage{
data_laterals
}
\description{
Data on the realization of laterals in "dark" (i.e. non-prevocalic) contexts by German learners of English, from a study by Pascoe (1987).
}
\keyword{datasets}
