% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_passives}
\alias{data_passives}
\title{The passive voice in academic writing}
\format{
\subsection{\code{data_passives}}{

A data frame with 160 rows and 6 columns:
\describe{
\item{id}{text ID}
\item{corpus}{source corpus ("Brown" vs. "Frown")}
\item{act}{number of active verb phrases in the text}
\item{pass}{number of passive verb phrases in the text}
\item{verbs}{total number of verb phrases in the text}
\item{prop_passive}{proportion of passive verb phrases in the text}
}
}
}
\source{
Evert, Stephanie. 2023. \emph{corpora: Statistics and Data Sets for Corpus Frequency Data}. doi:10.32614/CRAN.package.corpora \url{https://doi.org/10.32614/CRAN.package.corpora}, R package version 0.6, \url{https://CRAN.R-project.org/package=corpora}.

Francis, W. N. & H. Kučera. 1979. A Standard Corpus of Present-Day Edited American English, for Use with Digital Computers (Brown). Providence, RI: Brown University.

Mair, Christian. 1999. The Freiburg-Brown Corpus (‘Frown’). Original edition. Freiburg: Albert-Ludwigs-Universität.
}
\usage{
data_passives
}
\description{
Data on the frequency of the passive voice in American English academic writing, drawn from the Brown Corpus (Francis and Kučera 1979) and the Freiburg-Brown Corpus ('Frown'; Mair 1999). This dataset was originally made available as part of the R package \code{corpora} (Evert 2023).
}
\keyword{datasets}
