% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_should}
\alias{data_should}
\title{The frequency of "should" in written American English}
\format{
\subsection{\code{data_should}}{

A data frame with 1,000 rows and 6 columns:
\describe{
\item{text_id}{text ID}
\item{n_tokens}{number of occurrences of "should" in the text}
\item{corpus}{source corpus ("Brown" vs. "Frown")}
\item{genre}{broad genre ("Fiction", "General prose", "Learned", "Press")}
\item{subgenre}{subgenre}
\item{n_words}{length of the text (number of word tokens)}
}
}
}
\source{
Francis, W. N. & H. Kučera. 1979. A Standard Corpus of Present-Day Edited American English, for Use with Digital Computers (Brown). Providence, RI: Brown University.

Mair, Christian. 1999. The Freiburg-Brown Corpus (‘Frown’). Original edition. Freiburg: Albert-Ludwigs-Universität.

Sönning, Lukas. 2024. Background data for: Some obstacles to replication in corpus linguistics. DataverseNO, V1. https://doi.org/10.18710/7LNWJX
}
\usage{
data_should
}
\description{
Data on the frequency of the modal verb "should" in written American English, drawn from the Brown Corpus (Francis and Kučera 1979) and the Freiburg-Brown Corpus ('Frown'; Mair 1999). This is a subset from a larger dataset on the frequency of modal verbs in the Brown Family of corpora (Sönning 2023).
}
\keyword{datasets}
