% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_facet_nested.R
\name{theme_facet_nested}
\alias{theme_facet_nested}
\title{ggplot2 theme for nested facets}
\usage{
theme_facet_nested()
}
\description{
ggplot2 theme for drawing nested facets with the {ggh4x} package, in part based on the code provided by Teun Brand on stackoverflow.com (https://stackoverflow.com/questions/40316169/nested-facets-in-ggplot2-spanning-groups).
}
