% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_spike_graph.R
\name{theme_spike_graph}
\alias{theme_spike_graph}
\title{ggplot2 theme for spike graph with nested facets}
\usage{
theme_spike_graph()
}
\description{
ggplot2 theme for drawing nested facets with the {ggh4x} package, in part based on the code provided by Teun van den Brand on stackoverflow.com (https://stackoverflow.com/questions/40316169/nested-facets-in-ggplot2-spanning-groups).
}
